Force Abaqus's PMPI to not do any CPU affinity
The IBM Platform MPI bits are doing odd things with CPU affinity, resulting
in all the processes spawned by it being pinned to a single CPU.  I tried
various things trying to get PMPI to not do anything with affinity by
setting environment variables and command switches as advised in the manual
http://publibfp.dhe.ibm.com/epubs/pdf/c2753190.pdf.  Nothing worked.

In the end I wrote this small shared object library to override the
Linux sched_setaffinity() system call to return failure and errno = EINVAL
to keep PMPI from doing whatever it is doing... badly.

This will result in the processes of the job on a worker node having
a scheduling domain across all the CPUs on the node.  So a slight performance
impact will result if the kernel shuffles the processes around CPUs badly.
However, it's better than all the processes contending for a single CPU core.